Fixes the docker conductor:ui build.
There appears to be a weird error with the node dependencies and the current node version
used in the docker UI container image. You can find more information: https://github.com/gulpjs/gulp/issues/2162.

If you run the conductor UI container, it will not startup properly because the npm run build --server will fail with the following error:

```
../src/node_contextify.cc:631:static void node::contextify::ContextifyScript::New(const FunctionCallbackInfo<v8::Value> &): Assertion `args[1]->IsString()' failed.
```

Which is caused by the natives library. I updated the natives package directly in the lock file. I couldn't figure out the magical incantation for npm to do it for me.

This isn't caught in the Dockerfile for some reason. Even after using set -e, docker will run the command and continue on.
Not sure why that happens, but you can run the command from the specific layer to see the real error message.
Because the dist folder is never created, the startup.sh script will not run.