Update logic for processing of unacks
I believe the intent for "processUnacks" and "processAllUnacks" is to un-ack messages that have been lingering for longer than a minute (60 seconds is hardcoded in the query).  The problem is that the current logic adds that minute to the current time (which always results as a minute into the future) and tests if it is greater than the deliver_on time. This is always going to return true.  The desired effect would be garnered by SUBTRACTing the 60 seconds from current time, thus comparing deliver_on to 1 minute in the past to see if the minute has passed.

The result of the current logic is that all messages that are APPROACHING deliver_on time (within the next 60 seconds) are unacked every time one of these functions is called, regardless of whether deliver_on has been reached, let alone passed.