Make Conductor Flyway more configurable
When using Conductor with an existing schema that is also managed by Flyway, it is difficult to keep them from clashing. This is because the Conductor migrations are on the default Flyway locations path and Flyway uses Classpath scanning, so it will combine all migrations together leading to version clashes and out of order migration issues.

- Move Postgres and MySQL Flyway migrations from default Flyway location so as not to clash with external Flyway migrations
- Add flyway.locations property to make the locations externally configurable
- Add Flyway baseline migration config parameter to allow Conductor migration in schemas with existing tables
- Update Postgres and MySQL tests to utilize new default locations.