Make Event System Task Synchronous
Event is a task that publishes messages to an external eventing system.
The publishing is not a long running task, and publishing message to an async system using async task is not optimal.

marking event as sync, improves the overall performance and has negligible overhead given the task completes quite quick.