Task details not showing in UI
Event.path is deprecated in chrome as of the last update https://chromestatus.com/feature/5726124632965120. This is causing a bug when clicking on a task in the UI. This is also see in safari: https://github.com/Netflix/conductor/issues/3330. Updating to use composedPath() instead per guidance.

This is compatible with all browsers:
https://developer.mozilla.org/en-US/docs/Web/API/Event/composedPath